Move module-level functions back to 'web'
