+++ tornado/httputil.py
@@ -16,9 +16,6 @@
 
 """HTTP utility code shared by clients and servers."""
 
-import base64
-import hashlib
-import hmac
 import logging
 import re
 import time
@@ -257,55 +254,6 @@ def _parse_header(line):
                 value = value.replace('\\\\', '\\').replace('\\"', '"')
             pdict[name] = value
     return key, pdict
-
-
-def create_signed_value(secret, name, value):
-    timestamp = utf8(str(int(time.time())))
-    value = base64.b64encode(utf8(value))
-    signature = _create_signature(secret, name, value, timestamp)
-    return b("|").join([value, timestamp, signature])
-
-def decode_signed_value(secret, name, value, max_age=0):
-    parts = utf8(value).split(b("|"))
-    if len(parts) != 3: return None
-    value, timestamp, signature = parts
-    if not _time_independent_equals(signature, _create_signature(secret, name,
-                                                                 value,
-                                                                 timestamp)):
-        return None
-    if timestamp.startswith(b("0")):
-        return None
-    timestamp = int(timestamp)
-    if max_age > 0 and timestamp < time.time() - max_age:
-        return None
-    if timestamp > time.time() + 31 * 86400:
-        # _create_signature does not hash a delimiter between the
-        # parts of the cookie, so an attacker could transfer trailing
-        # digits from the payload to the timestamp without altering the
-        # signature.  For backwards compatibility, sanity-check timestamp
-        # here instead of modifying _cookie_signature.
-        return None
-    try:
-        return base64.b64decode(value)
-    except Exception:
-        return None
-
-def _create_signature(secret, *parts):
-    hash = hmac.new(utf8(secret), digestmod=hashlib.sha1)
-    for part in parts: hash.update(utf8(part))
-    return utf8(hash.hexdigest())
-
-def _time_independent_equals(a, b):
-    if len(a) != len(b):
-        return False
-    result = 0
-    if type(a[0]) is int:  # python3 byte strings
-        for x, y in zip(a,b):
-            result |= x ^ y
-    else:  # python2
-        for x, y in zip(a, b):
-            result |= ord(x) ^ ord(y)
-    return result == 0
 
 
 def doctests():
